requirements.txt
file-magic==0.4.0
